pyproject.toml
[tool.poetry]
name = "habitscoursework7"
version = "0.1.0"
description = ""
authors = ["Olga Lebedeva <olga.s.lebedeva@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
ipython = "^8.23.0"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
coverage = "^7.4.4"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
